Page refactoring
